refactor(textify): extract shared value primitives into neutral textify::values module
Move Value, NamedArg, Arguments, ArgsLayout, ValueEnum, enum_str_value, and the
SortField->Value conversion out of textify::rels into a new neutral
textify::values module, and add a shared decode_enum_field helper (dedups the
raw-i32-enum -> Value decode, applied here to SetRel). This decouples
textify::expressions from textify::rels: both now depend downward on
textify::values, which is a leaf with no back-dependency.

Behavior-preserving; no grammar or parser changes.
