+++ src/textify/expressions.rs
@@ -1,4 +1,4 @@
-use std::fmt::{self};
+use std::fmt;
 
 use chrono::{DateTime, NaiveDate};
 use expr::RexType;
@@ -402,32 +402,45 @@ impl Textify for FieldReference {
     }
 }
 
+/// Write the `name(args, options)` prefix shared by `ScalarFunction`,
+/// `AggregateFunction`, and `WindowFunction` textification.
+fn textify_function_call_prefix<S: Scope, W: fmt::Write>(
+    function_reference: u32,
+    arguments: &[FunctionArgument],
+    options: &[FunctionOption],
+    ctx: &S,
+    w: &mut W,
+) -> fmt::Result {
+    let name_and_anchor = NamedAnchor::lookup(ctx, ExtensionKind::Function, function_reference);
+    let name_and_anchor = ctx.display(&name_and_anchor);
+
+    let between = if arguments.is_empty() || options.is_empty() {
+        ""
+    } else {
+        ", "
+    };
+    let args = ctx.separated(arguments, ", ");
+    let options = ctx.separated(options, ", ");
+
+    write!(w, "{name_and_anchor}({args}{between}{options})")
+}
+
 impl Textify for ScalarFunction {
     fn name() -> &'static str {
         "ScalarFunction"
     }
 
     fn textify<S: Scope, W: fmt::Write>(&self, ctx: &S, w: &mut W) -> fmt::Result {
-        let name_and_anchor =
-            NamedAnchor::lookup(ctx, ExtensionKind::Function, self.function_reference);
-        let name_and_anchor = ctx.display(&name_and_anchor);
-
-        let args = ctx.separated(&self.arguments, ", ");
-        let options = ctx.separated(&self.options, ", ");
-        let between = if self.arguments.is_empty() || self.options.is_empty() {
-            ""
-        } else {
-            ", "
-        };
+        textify_function_call_prefix(
+            self.function_reference,
+            &self.arguments,
+            &self.options,
+            ctx,
+            w,
+        )?;
 
         let output = OutputType(self.output_type.as_ref());
-        let output_type = ctx.display(&output);
-
-        write!(
-            w,
-            "{name_and_anchor}({args}{between}{options}){output_type}"
-        )?;
-        Ok(())
+        write!(w, "{}", ctx.display(&output))
     }
 }
 
@@ -532,7 +545,7 @@ impl Textify for RexType {
             RexType::Literal(literal) => literal.textify(ctx, w),
             RexType::Selection(f) => f.textify(ctx, w),
             RexType::ScalarFunction(s) => s.textify(ctx, w),
-            RexType::WindowFunction(_w) => write!(
+            RexType::WindowFunction(_f) => write!(
                 w,
                 "{}",
                 ctx.failure(PlanError::unimplemented(
@@ -645,26 +658,16 @@ impl Textify for AggregateFunction {
     }
 
     fn textify<S: Scope, W: fmt::Write>(&self, ctx: &S, w: &mut W) -> fmt::Result {
-        // Similar to ScalarFunction textification
-        let name_and_anchor =
-            NamedAnchor::lookup(ctx, ExtensionKind::Function, self.function_reference);
-        let name_and_anchor = ctx.display(&name_and_anchor);
-
-        let args = ctx.separated(&self.arguments, ", ");
-        let options = ctx.separated(&self.options, ", ");
-        let between = if self.arguments.is_empty() || self.options.is_empty() {
-            ""
-        } else {
-            ", "
-        };
+        textify_function_call_prefix(
+            self.function_reference,
+            &self.arguments,
+            &self.options,
+            ctx,
+            w,
+        )?;
 
         let output = OutputType(self.output_type.as_ref());
-        let output_type = ctx.display(&output);
-
-        write!(
-            w,
-            "{name_and_anchor}({args}{between}{options}){output_type}"
-        )
+        write!(w, "{}", ctx.display(&output))
     }
 }
