+++ src/textify/rels.rs
@@ -2,28 +2,26 @@ use std::borrow::Cow;
 use std::collections::HashSet;
 use std::convert::TryFrom;
 use std::fmt;
-use std::fmt::Debug;
 
-use prost::{Message, UnknownEnumValue};
+use prost::Message;
 use substrait::proto::fetch_rel::CountMode;
 use substrait::proto::plan_rel::RelType as PlanRelType;
 use substrait::proto::read_rel::ReadType;
 use substrait::proto::rel::RelType;
 use substrait::proto::rel_common::EmitKind;
-use substrait::proto::sort_field::{SortDirection, SortKind};
 use substrait::proto::{
-    AggregateFunction, AggregateRel, CrossRel, Expression, ExtensionLeafRel, ExtensionMultiRel,
-    ExtensionSingleRel, FetchRel, FilterRel, JoinRel, NamedStruct, PlanRel, ProjectRel, ReadRel,
-    Rel, RelCommon, RelRoot, SetRel, SortField, SortRel, Type, join_rel, set_rel,
+    AggregateRel, CrossRel, ExtensionLeafRel, ExtensionMultiRel, ExtensionSingleRel, FetchRel,
+    FilterRel, JoinRel, NamedStruct, PlanRel, ProjectRel, ReadRel, Rel, RelCommon, RelRoot, SetRel,
+    SortRel, join_rel, set_rel,
 };
 
 use super::addenda::AddendumLines;
-use super::expressions::Reference;
 use super::types::Name;
+use super::values::{ArgsLayout, Arguments, NamedArg, Value, ValueEnum, decode_enum_field};
 use super::{PlanError, Scope, Textify};
 use crate::FormatError;
 use crate::extensions::any::AnyRef;
-use crate::extensions::{ExtensionArgs, ExtensionColumn, ExtensionError, ExtensionValue};
+use crate::extensions::{ExtensionArgs, ExtensionError};
 
 pub trait NamedRelation {
     fn name(&self) -> &'static str;
@@ -68,84 +66,6 @@ impl Textify for Rel {
         // delegates to `Relation` which carries `advanced_extension`, so the full
         // header → enhancement → children sequence is handled uniformly there.
         Relation::from_rel(self, ctx).textify(ctx, w)
-    }
-}
-
-/// Trait for enums that can be converted to a string representation for
-/// textification.
-///
-/// Returns Ok(str) for valid enum values, or Err([PlanError]) for invalid or
-/// unknown values.
-pub trait ValueEnum {
-    fn as_enum_str(&self) -> Result<Cow<'static, str>, PlanError>;
-}
-
-#[derive(Debug, Clone)]
-pub struct NamedArg<'a> {
-    pub name: Cow<'a, str>,
-    pub value: Value<'a>,
-}
-
-#[derive(Debug, Clone)]
-pub enum Value<'a> {
-    TableName(Vec<Name<'a>>),
-    Field(Option<Name<'a>>, Option<&'a Type>),
-    Tuple(Vec<Value<'a>>),
-    Reference(i32),
-    Expression(&'a Expression),
-    AggregateFunction(&'a AggregateFunction),
-    /// Represents a missing, invalid, or unspecified value.
-    Missing(PlanError),
-    /// Represents a valid enum value as a string for textification.
-    Enum(Cow<'a, str>),
-    EmptyGroup,
-    Integer(i64),
-    /// A decoded extension argument value.
-    ExtensionArgument(ExtensionValue),
-    /// A decoded extension output column.
-    ExtColumn(ExtensionColumn),
-}
-
-impl<'a> Value<'a> {
-    pub fn expect(maybe_value: Option<Self>, f: impl FnOnce() -> PlanError) -> Self {
-        match maybe_value {
-            Some(s) => s,
-            None => Value::Missing(f()),
-        }
-    }
-}
-
-impl<'a> From<Result<Vec<Name<'a>>, PlanError>> for Value<'a> {
-    fn from(token: Result<Vec<Name<'a>>, PlanError>) -> Self {
-        match token {
-            Ok(value) => Value::TableName(value),
-            Err(err) => Value::Missing(err),
-        }
-    }
-}
-
-impl<'a> Textify for Value<'a> {
-    fn name() -> &'static str {
-        "Value"
-    }
-
-    fn textify<S: Scope, W: fmt::Write>(&self, ctx: &S, w: &mut W) -> fmt::Result {
-        match self {
-            Value::TableName(names) => write!(w, "{}", ctx.separated(names, ".")),
-            Value::Field(name, typ) => {
-                write!(w, "{}:{}", ctx.expect(name.as_ref()), ctx.expect(*typ))
-            }
-            Value::Tuple(values) => write!(w, "({})", ctx.separated(values, ", ")),
-            Value::Reference(i) => write!(w, "{}", Reference(*i)),
-            Value::Expression(e) => write!(w, "{}", ctx.display(*e)),
-            Value::AggregateFunction(agg_fn) => agg_fn.textify(ctx, w),
-            Value::Missing(err) => write!(w, "{}", ctx.failure(err.clone())),
-            Value::Enum(res) => write!(w, "&{res}"),
-            Value::Integer(i) => write!(w, "{i}"),
-            Value::EmptyGroup => write!(w, "_"),
-            Value::ExtensionArgument(ev) => ev.textify(ctx, w),
-            Value::ExtColumn(ec) => ec.textify(ctx, w),
-        }
     }
 }
 
@@ -295,68 +215,6 @@ impl<'a> Textify for Emitted<'a> {
     }
 }
 
-/// How an argument list renders inside a relation's `[...]`.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub enum ArgsLayout {
-    /// `arg, arg, arg` on a single line.
-    #[default]
-    Inline,
-    /// One `- arg` per line, used for `Read:Virtual` rows. See
-    /// [`Relation::write_header`] for the exact layout.
-    Rows,
-}
-
-#[derive(Debug, Clone)]
-pub struct Arguments<'a> {
-    /// Positional arguments (e.g., a filter condition, group-bys, etc.)
-    pub positional: Vec<Value<'a>>,
-    /// Named arguments (e.g., limit=10, offset=5)
-    pub named: Vec<NamedArg<'a>>,
-    /// How this argument list is laid out. Defaults to [`ArgsLayout::Inline`];
-    /// only `Read:Virtual` opts into [`ArgsLayout::Rows`].
-    layout: ArgsLayout,
-}
-
-impl<'a> Arguments<'a> {
-    /// An inline argument list (`arg, arg, arg`), the default for every
-    /// relation.
-    pub fn inline(positional: Vec<Value<'a>>, named: Vec<NamedArg<'a>>) -> Self {
-        Arguments {
-            positional,
-            named,
-            layout: ArgsLayout::Inline,
-        }
-    }
-
-    /// A row-per-line argument list (`- arg` per line) used for `Read:Virtual`
-    /// with many rows. Currently not enabled for named arguments.
-    /// TODO: enable for named arguments as well.
-    pub fn rows(positional: Vec<Value<'a>>) -> Self {
-        Arguments {
-            positional,
-            named: vec![],
-            layout: ArgsLayout::Rows,
-        }
-    }
-}
-
-impl<'a> Textify for Arguments<'a> {
-    fn name() -> &'static str {
-        "Arguments"
-    }
-    fn textify<S: Scope, W: fmt::Write>(&self, ctx: &S, w: &mut W) -> fmt::Result {
-        if self.positional.is_empty() && self.named.is_empty() {
-            return write!(w, "_");
-        }
-
-        write!(w, "{}", ctx.separated(self.positional.iter(), ", "))?;
-        if !self.positional.is_empty() && !self.named.is_empty() {
-            write!(w, ", ")?;
-        }
-        write!(w, "{}", ctx.separated(self.named.iter(), ", "))
-    }
-}
-
 pub struct Relation<'a> {
     pub name: Cow<'a, str>,
     /// Arguments to the relation, if any.
@@ -423,7 +281,7 @@ impl Relation<'_> {
                 let cols = ctx.display(&cols);
                 write!(w, "{indent}{name}[{cols}]")
             }
-            Some(args) if args.layout == ArgsLayout::Rows => {
+            Some(args) if args.layout() == ArgsLayout::Rows => {
                 // One `- row` per line, one indent level deeper, with a trailing
                 // comma on every row but the last, then `- <output> cols]`.
                 let child = ctx.push_indent();
@@ -1155,17 +1013,7 @@ impl<'a> Relation<'a> {
             total_columns / children.len()
         };
 
-        let op_value = match set_rel::SetOp::try_from(rel.op) {
-            Ok(op) => match op.as_enum_str() {
-                Ok(s) => Value::Enum(s),
-                Err(e) => Value::Missing(e),
-            },
-            Err(_) => Value::Missing(PlanError::invalid(
-                "SetRel",
-                Some("op"),
-                format!("Unknown set op: {}", rel.op),
-            )),
-        };
+        let op_value = decode_enum_field::<set_rel::SetOp>(rel.op, "SetRel", "op");
 
         let arguments = Some(Arguments::inline(vec![op_value], vec![]));
         let emit = get_emit(rel.common.as_ref());
@@ -1204,147 +1052,6 @@ impl<'a> Relation<'a> {
     }
 }
 
-impl<'a> From<&'a SortField> for Value<'a> {
-    fn from(sf: &'a SortField) -> Self {
-        let field = match &sf.expr {
-            Some(expr) => match &expr.rex_type {
-                Some(substrait::proto::expression::RexType::Selection(fref)) => {
-                    if let Some(substrait::proto::expression::field_reference::ReferenceType::DirectReference(seg)) = &fref.reference_type {
-                        if let Some(substrait::proto::expression::reference_segment::ReferenceType::StructField(sf)) = &seg.reference_type {
-                            Value::Reference(sf.field)
-                        } else { Value::Missing(PlanError::unimplemented("SortField", Some("expr"), "Not a struct field")) }
-                    } else { Value::Missing(PlanError::unimplemented("SortField", Some("expr"), "Not a direct reference")) }
-                }
-                _ => Value::Missing(PlanError::unimplemented(
-                    "SortField",
-                    Some("expr"),
-                    "Not a selection",
-                )),
-            },
-            None => Value::Missing(PlanError::unimplemented(
-                "SortField",
-                Some("expr"),
-                "Missing expr",
-            )),
-        };
-        let direction = match &sf.sort_kind {
-            Some(kind) => Value::from(kind),
-            None => Value::Missing(PlanError::invalid(
-                "SortKind",
-                Some(Cow::Borrowed("sort_kind")),
-                "Missing sort_kind",
-            )),
-        };
-        Value::Tuple(vec![field, direction])
-    }
-}
-
-impl<'a, T: ValueEnum + ?Sized> From<&'a T> for Value<'a> {
-    fn from(enum_val: &'a T) -> Self {
-        match enum_val.as_enum_str() {
-            Ok(s) => Value::Enum(s),
-            Err(e) => Value::Missing(e),
-        }
-    }
-}
-
-impl ValueEnum for SortKind {
-    fn as_enum_str(&self) -> Result<Cow<'static, str>, PlanError> {
-        let d = match self {
-            &SortKind::Direction(d) => SortDirection::try_from(d),
-            SortKind::ComparisonFunctionReference(f) => {
-                return Err(PlanError::invalid(
-                    "SortKind",
-                    Some(Cow::Owned(format!("function reference{f}"))),
-                    "SortKind::ComparisonFunctionReference unimplemented",
-                ));
-            }
-        };
-        let s = match d {
-            Err(UnknownEnumValue(d)) => {
-                return Err(PlanError::invalid(
-                    "SortKind",
-                    Some(Cow::Owned(format!("unknown variant: {d:?}"))),
-                    "Unknown SortDirection",
-                ));
-            }
-            Ok(SortDirection::AscNullsFirst) => "AscNullsFirst",
-            Ok(SortDirection::AscNullsLast) => "AscNullsLast",
-            Ok(SortDirection::DescNullsFirst) => "DescNullsFirst",
-            Ok(SortDirection::DescNullsLast) => "DescNullsLast",
-            Ok(SortDirection::Clustered) => "Clustered",
-            Ok(SortDirection::Unspecified) => {
-                return Err(PlanError::invalid(
-                    "SortKind",
-                    Option::<Cow<str>>::None,
-                    "Unspecified SortDirection",
-                ));
-            }
-        };
-        Ok(Cow::Borrowed(s))
-    }
-}
-
-impl ValueEnum for join_rel::JoinType {
-    fn as_enum_str(&self) -> Result<Cow<'static, str>, PlanError> {
-        let s = match self {
-            join_rel::JoinType::Unspecified => {
-                return Err(PlanError::invalid(
-                    "JoinType",
-                    Option::<Cow<str>>::None,
-                    "Unspecified JoinType",
-                ));
-            }
-            join_rel::JoinType::Inner => "Inner",
-            join_rel::JoinType::Outer => "Outer",
-            join_rel::JoinType::Left => "Left",
-            join_rel::JoinType::Right => "Right",
-            join_rel::JoinType::LeftSemi => "LeftSemi",
-            join_rel::JoinType::RightSemi => "RightSemi",
-            join_rel::JoinType::LeftAnti => "LeftAnti",
-            join_rel::JoinType::RightAnti => "RightAnti",
-            join_rel::JoinType::LeftSingle => "LeftSingle",
-            join_rel::JoinType::RightSingle => "RightSingle",
-            join_rel::JoinType::LeftMark => "LeftMark",
-            join_rel::JoinType::RightMark => "RightMark",
-        };
-        Ok(Cow::Borrowed(s))
-    }
-}
-
-impl ValueEnum for set_rel::SetOp {
-    fn as_enum_str(&self) -> Result<Cow<'static, str>, PlanError> {
-        let s = match self {
-            set_rel::SetOp::Unspecified => {
-                return Err(PlanError::invalid(
-                    "SetOp",
-                    Option::<Cow<str>>::None,
-                    "Unspecified SetOp",
-                ));
-            }
-            set_rel::SetOp::MinusPrimary => "MinusPrimary",
-            set_rel::SetOp::MinusPrimaryAll => "MinusPrimaryAll",
-            set_rel::SetOp::MinusMultiset => "MinusMultiset",
-            set_rel::SetOp::IntersectionPrimary => "IntersectionPrimary",
-            set_rel::SetOp::IntersectionMultiset => "IntersectionMultiset",
-            set_rel::SetOp::IntersectionMultisetAll => "IntersectionMultisetAll",
-            set_rel::SetOp::UnionDistinct => "UnionDistinct",
-            set_rel::SetOp::UnionAll => "UnionAll",
-        };
-        Ok(Cow::Borrowed(s))
-    }
-}
-
-impl<'a> Textify for NamedArg<'a> {
-    fn name() -> &'static str {
-        "NamedArg"
-    }
-    fn textify<S: Scope, W: fmt::Write>(&self, ctx: &S, w: &mut W) -> fmt::Result {
-        write!(w, "{}=", self.name)?;
-        self.value.textify(ctx, w)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use substrait::proto::aggregate_rel::Grouping;
@@ -1355,12 +1062,13 @@ mod tests {
     use substrait::proto::rel_common::{Direct, Emit};
     use substrait::proto::r#type::{self as ptype, Boolean, I64, Kind, Nullability, Struct};
     use substrait::proto::{
-        Expression, FunctionArgument, NamedStruct, ReadRel, Type, aggregate_rel,
+        AggregateFunction, Expression, FunctionArgument, NamedStruct, ReadRel, Type, aggregate_rel,
     };
 
     use super::*;
     use crate::fixtures::TestContext;
     use crate::parser::expressions::FieldIndex;
+    use crate::textify::expressions::Reference;
 
     #[test]
     fn test_read_rel() {
